tcp: mitigate illegal state transitions on simultaneous close
If a socket is closed and the remote side sends a RST before the local
side sends its own RST, it is possible to jump directly from FIN-WAIT-1
to CLOSING without ever sending a RST to the remote side.

The main side effect is that if untransmitted data is present, the
socket is never exhausted, causing `EthernetInterface::poll` to loop
forever.

The ideal fix is to follow RFC more closely and not transition to
FIN-WAIT-1 until the RST has been sent, but there are too many
assumptions based on the current state that would be broken by remaining
in ESTABLISHED with a closed transmit buffer to be worth fixing
a relatively rare edge case. Instead, add another illegal transition to
fix the failures of a previous violation, in the spirit of all good TCP
stacks.